+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
+# [0.6.0](https://github.com/johnf1004/google_cloud_utilities/compare/v0.5.0...v0.6.0) (2023-12-14)
+
+### Features
+
+* New function append_rows_bq_pandas for smoother way to insert rows to table with pandas (with labels in job config)
+
 # [0.5.0](https://github.com/johnf1004/google_cloud_utilities/compare/v0.4.4...v0.5.0) (2023-12-13)
 
 ### Features
 
-* New function append_rows_bq_json for smoother way to insert rows to table (with labels in job config)
+* New function append_rows_bq_json for smoother way to insert rows to table with dictionaries (with labels in job config)
 
 # [0.4.4](https://github.com/johnf1004/google_cloud_utilities/compare/v0.4.3...v0.4.4) (2023-12-12)